fix(ble): 修复蓝牙GATT特征索引和串口发送函数参数命名问题
- 修正BLE通知发送时使用的GATT属性索引从attrs[3]改为attrs[4]
- 更新Android端蓝牙服务中TX和RX特征UUID的注释说明和对应关系
- 将串口发送函数SendData的参数名length统一改为len以保持一致性
- 在日志输出中为控制数据坐标值添加显式类型转换以避免潜在类型问题
- 更新应用版本号从0.4.0到0.5.0

+++ firmware/zephyr_app/app/src/main.cpp
@@ -39,7 +39,7 @@ static void OnControlDataReceived(const ControlData& data) {
 		// TODO: 执行急停逻辑
 	} else {
 		LOG_INF("Control data: L=(%d,%d) R=(%d,%d) buttons=0x%08X",
-				data.left_x, data.left_y, data.right_x, data.right_y, data.buttons);
+				(int)data.left_x, (int)data.left_y, (int)data.right_x, (int)data.right_y, (int)data.buttons);
 	}
 
 	// 转发控制数据到 MCU (通过 UART)
